make SPDK installation step optional (#1)
Omit installation of SPDK to support projects that do not use it.

+++ main.sh
@@ -2,15 +2,23 @@
 set -e
 set -o pipefail
 
-CODEROOT=$HOME/setup-dpdk
-mkdir -p $CODEROOT/dpdk_$DPDKVER $CODEROOT/spdk_$SPDKVER
-
-cd $CODEROOT/spdk_$SPDKVER
-if ! [[ -f scripts/pkgdep.sh ]]; then
-  curl -sfL https://github.com/spdk/spdk/archive/v$SPDKVER.tar.gz | tar -xz --strip-components=1
+# Install all dependencies based on the OS distribution
+if [ -n "$(command -v apt-get)" ]; then
+  sudo apt-get update;
+  sudo apt-get -y install \
+    jq \
+    ninja-build \
+    python3 \
+    python3-pip \
+    python3-pyelftools \
+    python3-setuptools \
+    python3-wheel
+    
+  sudo pip3 install meson
 fi
-sudo apt-get install python3-pyelftools python3-setuptools
-sudo scripts/pkgdep.sh
+
+CODEROOT=$HOME/setup-dpdk
+mkdir -p $CODEROOT/dpdk_$DPDKVER
 
 cd $CODEROOT/dpdk_$DPDKVER
 if ! [[ -f meson.build ]]; then
@@ -29,15 +37,25 @@ sudo ninja -C build install
 sudo find /usr/local/lib -name 'librte_*.a' -delete
 sudo ldconfig
 
-cd $CODEROOT/spdk_$SPDKVER
-if ! [[ -f build/lib/libspdk_env_dpdk.a ]]; then
-  ./configure --target-arch=$TARGETARCH --enable-debug --disable-tests --with-shared \
-    --with-dpdk=/usr/local --without-vhost --without-isal --without-fuse
-  make -j$(nproc)
+if ! [[ -z $SPDKVER ]] && [[ -n $SPDKVER ]] && [[ $SPDKVER != "none" ]]; then
+  mkdir -p  $CODEROOT/spdk_$SPDKVER
+  cd $CODEROOT/spdk_$SPDKVER
+  if ! [[ -f scripts/pkgdep.sh ]]; then
+    curl -sfL https://github.com/spdk/spdk/archive/v$SPDKVER.tar.gz | tar -xz --strip-components=1
+  fi
+
+  sudo scripts/pkgdep.sh
+
+  cd $CODEROOT/spdk_$SPDKVER
+  if ! [[ -f build/lib/libspdk_env_dpdk.a ]]; then
+    ./configure --target-arch=$TARGETARCH --enable-debug --disable-tests --with-shared \
+      --with-dpdk=/usr/local --without-vhost --without-isal --without-fuse
+    make -j$(nproc)
+  fi
+  sudo make install
+  sudo find /usr/local/lib -name 'libspdk_*.a' -delete
+  sudo ldconfig
 fi
-sudo make install
-sudo find /usr/local/lib -name 'libspdk_*.a' -delete
-sudo ldconfig
 
 if [[ $NRHUGE -gt 0 ]]; then
   echo $NRHUGE | sudo tee /sys/devices/system/node/node*/hugepages/hugepages-2048kB/nr_hugepages >/dev/null
